added delays for downloading geyser and spigot

+++ run.js
@@ -37,12 +37,14 @@ if (!fs.existsSync("servers.json")) {
   fs.writeFileSync("servers.json", "{}");
 }
 const s = require("./scripts/stripe.js");
-exec(
-  "curl -o servers/template/cx_geyser-spigot_Geyser.jar -LO https://ci.opencollab.dev/job/GeyserMC/job/Geyser/job/master/lastSuccessfulBuild/artifact/bootstrap/spigot/build/libs/Geyser-Spigot.jar"
-);
-exec(
-  "curl -o servers/template/cx_floodgate-spigot_Floodgate.jar -LO https://ci.opencollab.dev/job/GeyserMC/job/Floodgate/job/master/lastSuccessfulBuild/artifact/spigot/build/libs/floodgate-spigot.jar"
-);
+if (!fs.existsSync("servers/template/cx_geyser-spigot_Geyser.jar")) {
+  exec(
+    "curl -o servers/template/cx_geyser-spigot_Geyser.jar -LO https://ci.opencollab.dev/job/GeyserMC/job/Geyser/job/master/lastSuccessfulBuild/artifact/bootstrap/spigot/build/libs/Geyser-Spigot.jar"
+  );
+  exec(
+    "curl -o servers/template/cx_floodgate-spigot_Floodgate.jar -LO https://ci.opencollab.dev/job/GeyserMC/job/Floodgate/job/master/lastSuccessfulBuild/artifact/spigot/build/libs/floodgate-spigot.jar"
+  );
+}
 if (!fs.existsSync("java")) {
   fs.mkdirSync("java");
